Improve code file structure
