+++ Ecom/components/drawer/drawerContents.js
@@ -1,0 +1,37 @@
+export const drawerContents = [
+  {
+    labal: "Profile",
+    icon: "ios-person-circle-outline",
+    iconLibrary: "Ionicons",
+    pathName: "/profile",
+    route: "/(tabs)/profile",
+  },
+  {
+    labal: "Refer & Earn Program",
+    icon: "people-circle-outline",
+    iconLibrary: "Ionicons",
+    pathName: "/referEarn",
+    route: "/referEarn",
+  },
+  {
+    labal: "Points",
+    icon: "stars",
+    iconLibrary: "MaterialIcons",
+    pathName: "/points",
+    route: "/points",
+  },
+  {
+    labal: "Tutorials",
+    icon: "video",
+    iconLibrary: "Feather",
+    pathName: "/tutorials",
+    route: "/tutorials",
+  },
+  {
+    labal: "Logout",
+    icon: "logout",
+    iconLibrary: "MaterialIcons",
+    pathName: "/logout",
+    route: "/(tabs)/profile",
+  },
+];
