+++ Ecom/components/index.js
@@ -4,3 +4,5 @@ export { StyledInput } from "./inputs/StyledInput";
 export { StyledButton } from "./buttons/StyledButton";
 //texts
 export { StyledText } from "./texts/StyledText";
+//drawer
+export { CustomDrawerContent } from "./drawer/CustomDrawerContent";
